[Chore #160538126] restructures the dirs  (versioning and blueprints)
