+++ Pipfile
@@ -15,6 +15,7 @@ coverage = "*"
 coveralls = "*"
 gunicorn = "*"
 python-dotenv = "*"
+"psycopg2" = "*"
 
 [dev-packages]
