+++ app/__init__.py
@@ -1,13 +1,19 @@
 """ Initialization file """
-
+import os
 from flask import Flask
-
 
 # Init the app
 app = Flask(__name__, instance_relative_config=True)
 
-# load views
-from app import views
 
 # Load the config file
-app.config.from_object('config')
+app.config.from_object('instance.config')
+
+
+# load views
+from app.api.v1 import views
+
+app.register_blueprint(views.v1_bp)
+
+ 
+
