+++ instance/config.py
@@ -1,4 +1,4 @@
 """ Project configuration variables """
-
 # Enable Debugging
+
 DEBUG = True
