+++ run.py
@@ -1,5 +1,6 @@
 """ Application root. Run first to start server and launch app """
 
+import os
 from app import app
